feat: Add ability to set child groups
Add function to call the `POST /{realm}/groups/{id}/children` API endpoint
in order to create or update a group as a child of another. This function
checks beforehand that the group isn't already a child, as the API will
respond with an error if it already is.

Modify the `FindGroupByName` function to search recursively within
through the group tree.