feat: Add ability to find group by path
The current FindGroupByName function will find a group by name "bar" in
the following hierarchy:
```
. foo
└── bar
```

However, there is no functionality for searching for the exact path "foo/bar".
Add a `FindGroupByPath` function to the client that traverses group tree
using the given path until it either finds the exact group, or returns
`nil` if the path doesn't match the tree.